fix: get merkle proof API and create a dao service for merkle proof
